Key down bug fixed
When navigating using left/right and then pressing down lead to a bug
where the down arror started to animate the same page when the expected
behaviour would be to switch to the next slide.

This is because both -1 and the max_value for lines on the slide mean
the same thing, show the whole slide. This should be fixed in the future
I think but for now this hack works fine.