Exception handling to if text does not fit window
When using Windows Terminal for example it is possible to set any font
size and there it is easy to "zoom" in too much which caused an crash
before.

Now that exception is handled and an error message is shown. It is a bit
lazy not checking for any particular exception but only curses.exception
in generall, but it is fine for now.